+++ example/lib/pages/enemy/enemy_page.dart
@@ -12,7 +12,7 @@ class EnemyPage extends StatelessWidget {
     const tileSize = 16.0;
     return BonfireWidget(
       map: WorldMapByTiled(
-        'tiled/punnyworld/simple_map.tmj',
+        TiledReader.asset('tiled/punnyworld/simple_map.tmj'),
         objectsBuilder: {
           'range': (properties) => RageEnemy(position: properties.position),
           'melee': (properties) => MeleeEnemy(position: properties.position),