+++ example/lib/pages/path_finding/path_finding_page.dart
@@ -10,7 +10,7 @@ class PathFindingPage extends StatelessWidget {
     const tileSize = 16.0;
     return BonfireWidget(
       map: WorldMapByTiled(
-        'tiled/punnyworld/pathfinding_map.tmj',
+        TiledReader.asset('tiled/punnyworld/pathfinding_map.tmj'),
       ),
       player: HumanPathFinding(
         position: Vector2(tileSize * 3, tileSize * 3),