+++ lib/tiled/cache_provider/tiled_cache_provider.dart
@@ -1,0 +1,5 @@
+abstract class TiledCacheProvider {
+  Future<Map<String, dynamic>> get(String key);
+  void put(String key, Map<String, dynamic> data);
+  Future<bool> containsKey(String key);
+}